fix(rubocop): split generator instrumentation event modules
Extract generation and scoring event helpers for label and self-text
instrumentation so ModuleLength can be re-enabled later.

Issue: No ticket/issue
